Login interfaces and types moved to file
