+++ frontend/src/Components/LoginPage/GoogleAuth.tsx
@@ -1,0 +1,7 @@
+export const GoogleAuthenticaion = () : JSX.Element => {
+	return (
+		<div className="container-fluid">
+
+		</div>
+	);
+};
